requirements.txt
altair==5.5.0
altgraph==0.17.4
async-timeout==5.0.1
attrs==25.3.0
black==24.10.0
blinker==1.9.0
business-rules-enhanced==1.4.8
cachetools==6.1.0
cdisc-library-client==0.1.6
Cerberus==1.3.7
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coverage==7.10.2
cramjam==2.11.0
dask==2024.6.0
dask-expr==1.1.3
distlib==0.4.0
elementpath==5.0.3
et_xmlfile==2.0.0
fastparquet==2024.2.0
filelock==3.18.0
flake8==6.1.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
identify==2.6.12
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
Jinja2==3.1.6
jsonpath-ng==1.6.1
jsonschema==4.18.5
jsonschema-specifications==2025.4.1
locket==1.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
narwhals==2.0.1
nodeenv==1.9.1
numpy==1.26.4
odmlib==0.1.4
openpyxl==3.1.5
packaging==25.0
pandas==2.1.4
partd==1.4.2
pathspec==0.12.1
pathvalidate==3.3.1
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
pre-commit==2.20.0
protobuf==6.31.1
psutil==6.1.1
pyarrow==21.0.0
pycodestyle==2.11.1
pydeck==0.9.1
pyflakes==3.1.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2025.8
Pympler==1.1
pyreadstat==1.2.7
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==4.5.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
virtualenv==20.33.1
watchdog==6.0.0
xmlschema==4.1.0
zipp==3.23.0
